iot_edge_ids/edr/requirements.txt
scikit-learn==1.4.1.post1
numpy==1.26.4
paho-mqtt==1.6.1
influxdb-client[ciso]==1.40.0
